fix!: file-scoped namespace and move enums to DisCatSharp.Enums

+++ DisCatSharp/Enums/Invite/InviteFlags.cs
@@ -22,7 +22,7 @@
 
 namespace DisCatSharp.Enums;
 
-public enum InviteFlags : long
+public enum InviteFlags : int
 {
 	/// <summary>
 	/// Invite has no flags
